+++ R/box_usage_helper_functions.R
@@ -122,7 +122,9 @@ get_special_calls <- function(xml) {
   )
 }
 
-#' Get objects called in current source file
+#' Get objects called in current source file.
+#'
+#' This ignores objects to the left of `<-`, `=`, `%<-%` as these are assignments.
 #'
 #' @param xml An XML node list
 #' @return a list of `xml_nodes` and `text`.
@@ -133,7 +135,9 @@ get_object_calls <- function(xml) {
     ./SYMBOL and
     not(
       following-sibling::LEFT_ASSIGN or
-      following-sibling::EQ_ASSIGN
+      following-sibling::EQ_ASSIGN or
+      parent::expr[
+        following-sibling::SPECIAL[text() = '%<-%']
     ) and
     not(
       ancestor::expr/expr[
@@ -171,4 +175,25 @@ get_function_signature_objs <- function(xml) {
 internal_r6_refs <- function(func_list) {
   r6_refs <- "self|private\\$.+"
   grepl(r6_refs, func_list)
+}
+
+#' Get the output object names of the deconstructor (`rhino::%<-%`) assignment operator.
+#'
+#' @description
+#' This is a naive search for the `SYMBOLS` within a `c()` as the first expression before
+#' the `%<-%`. For example: `c(x, y, z) %<-% ...`.
+#'
+#' @param xml An XML node list
+#' @return a list of `xml_nodes` and `text`
+#' @keywords internal
+get_deconstructor_objects <- function(xml) {
+  xpath_deconstructor_objects <- "
+  //expr[
+    SPECIAL[text() = '%<-%']
+  ]
+  /expr[1 and ./expr/SYMBOL_FUNCTION_CALL[text() = 'c']]
+  //SYMBOL
+  "
+
+  extract_xml_and_text(xml, xpath_deconstructor_objects)
 }