+++ NEWS.md
@@ -1,3 +1,7 @@
+# development version
+
+* Support for destructure operator `%<-%`
+
 # box.linters 0.10.5
 
 * Fix for `treesitter.r` update to version 1.1.0. Change in how treesitter returns the start row of the program node. (#143)